chore(redis): Moved cacher to own crate
