+++ src/services/redis_cache/values.rs
@@ -1,5 +1,5 @@
 use wrappers::document::Document;
-use wrappers::search_params::SearchParams;
+use wrappers::s_params::SearchParams;
 
 use redis::{ErrorKind, RedisError, RedisResult, Value};
 use redis::{FromRedisValue, RedisWrite, ToRedisArgs};
@@ -68,11 +68,11 @@ impl FromRedisValue for CacherDocuments {
         match value {
             Value::Data(data) => serde_json::from_slice::<CacherDocuments>(data.as_slice())
                 .map_err(|_| {
-                    let msg = "Failed while deserializing document from redis";
+                    let msg = "Failed while deserializing document from redis_cache";
                     RedisError::from((ErrorKind::IoError, msg))
                 }),
             _ => {
-                let err = anyhow::Error::msg("Incorrect redis value type to deserialize");
+                let err = anyhow::Error::msg("Incorrect redis_cache value type to deserialize");
                 let io_err = std::io::Error::new(std::io::ErrorKind::InvalidData, err);
                 Err(io_err.into())
             }
